[#178] Refactored Classes into Main Package
This moves the previously comitted packages all into the application's main project creating a better view where the activities can now be viewed as seperate from the other pieces of code.
