+++ app/src/androidTest/java/com/example/scoutconcordia/SwitchCampus.java
@@ -11,6 +11,8 @@ import androidx.test.rule.ActivityTestRule;
 import androidx.test.rule.GrantPermissionRule;
 import androidx.test.runner.AndroidJUnit4;
 
+import com.example.scoutconcordia.Activities.MapsActivity;
+
 import org.hamcrest.Description;
 import org.hamcrest.Matcher;
 import org.hamcrest.TypeSafeMatcher;
