+++ app/src/main/java/com/example/scoutconcordia/Activities/MainActivity.java
@@ -1,8 +1,10 @@
-package com.example.scoutconcordia;
+package com.example.scoutconcordia.Activities;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
+
+import com.example.scoutconcordia.R;
 
 public class MainActivity extends AppCompatActivity {
