+++ app/src/main/java/com/example/scoutconcordia/Activities/ShuttleScheduleActivity.java
@@ -1,10 +1,12 @@
-package com.example.scoutconcordia;
+package com.example.scoutconcordia.Activities;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.MenuItem;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.scoutconcordia.R;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 
 public class ShuttleScheduleActivity extends AppCompatActivity {
