+++ app/src/main/java/com/example/scoutconcordia/MapInfoClasses/BuildingInfo.java
