+++ app/src/main/res/layout/activity_calendar.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CalendarActivity">
+    tools:context=".Activities.CalendarActivity">
     <ScrollView
         android:layout_width="fill_parent"
         android:layout_height="fill_parent"
