+++ app/src/main/res/layout/activity_maps.xml
@@ -7,7 +7,7 @@ xmlns:tools="http://schemas.android.com/tools"
 android:name="com.google.android.gms.maps.SupportMapFragment"
 android:layout_width="match_parent"
 android:layout_height="match_parent"
-tools:context=".MapsActivity">
+tools:context=".Activities.MapsActivity">
 
 <fragment android:id="@+id/autocomplete_fragment"
         android:layout_width="match_parent"
@@ -35,7 +35,7 @@ tools:context=".MapsActivity">
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_below="@id/autocomplete_fragment"
-	tools:context=".MapsActivity"/>
+	tools:context=".Activities.MapsActivity"/>
 
 	<ToggleButton
         android:id="@+id/toggleButton"
